Initial commit of issue 2081
